Remove unnecessary code in BlockTicker and rename eventsRealyClass
